+++ smart-contracts/contracts/CosmosBridgeStorage.sol
@@ -71,7 +71,7 @@ contract CosmosBridgeStorage {
     /**
     * @notice network descriptor
     */
-    uint256 public networkDescriptor;
+    int32 public networkDescriptor;
 
     /**
     * @dev gap of storage for future upgrades
